add tracker

+++ internal/models/subscription.go
@@ -3,7 +3,7 @@ package models
 import "github.com/shopspring/decimal"
 
 type Subscription struct {
-	UserID uint `db:"user_id"`
-	ProductID uint `db:"product_id"`
+	UserID      int64           `db:"user_id"`
+	ProductID   int             `db:"product_id"`
 	TargetPrice decimal.Decimal `db:"target_price"`
 }
